requirements.txt
ultralytics
opencv-python
psycopg2-binary
streamlit
pandas
scikit-learn
notebook
lapx
boto3
yt-dlp
seaborn
minio